refactor: hook for ticket form
